requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# LangChain and LangGraph
langchain>=0.1.16
langgraph>=0.0.52

# OpenAI (for OpenRouter)
openai>=1.0.0

# Database (if needed)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# OCR and Document Processing
pytesseract>=0.3.10
Pillow>=10.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0

# Supabase Storage
supabase==2.16.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.24.0
